requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.21
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg[binary]
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
